Minor changes in testing and adding of Test Firebase
