+++ src/data/firebase/firebase.js
@@ -1,15 +1,7 @@
 // @flow
 
-const firebase = require('firebase');
-
-const config = {
-  apiKey: 'AIzaSyCB8YV-RAEjpwxjixq-stzkSc_JZgQw-88',
-  authDomain: 'pinup-cfc23.firebaseapp.com',
-  databaseURL: 'https://pinup-cfc23.firebaseio.com',
-  projectId: 'pinup-cfc23',
-  storageBucket: 'pinup-cfc23.appspot.com',
-  messagingSenderId: '522038590744',
-};
+import * as firebase from 'firebase';
+import config from './firebaseconfig';
 
 if (!firebase.apps.length) {
   firebase.initializeApp(config);
