+++ src/data/firebase/firebaseconfig.js
@@ -1,0 +1,11 @@
+// @flow
+
+const config = {
+  apiKey: 'AIzaSyCB8YV-RAEjpwxjixq-stzkSc_JZgQw-88',
+  authDomain: 'pinup-cfc23.firebaseapp.com',
+  databaseURL: 'https://pinup-cfc23.firebaseio.com',
+  projectId: 'pinup-cfc23',
+  storageBucket: 'pinup-cfc23.appspot.com',
+  messagingSenderId: '522038590744',
+};
+export default config;
